Refactor validation tests to use dynamic imports and add edge case tests
- Removed the CommonJS setup for TypeScript in validation tests.
- Converted validation tests to use async/await with dynamic imports.
- Added new edge case tests for diagram and palette validation.
- Consolidated test structure for better readability and maintainability.
- Deleted the old ES module validation test file.
- Introduced a new test file for transformation apply and export tests, covering various scenarios including internal wiring and duplicate port names.
