+++ test/run-tests.cjs
@@ -9,19 +9,8 @@ const path = require('path');
 async function runTest(testFile, testName) {
     return new Promise((resolve) => {
         console.log(`📋 Running ${testName}...`);
-        const env = { ...process.env };
-        const args = testFile.endsWith('.mjs')
-            ? ['--loader', 'ts-node/esm', testFile]
-            : [testFile];
-        if (testFile.endsWith('.mjs')) {
-            env.TS_NODE_PROJECT = path.resolve(__dirname, '../tsconfig.app.json');
-            env.TS_NODE_COMPILER_OPTIONS = JSON.stringify({
-                module: 'ESNext',
-                moduleResolution: 'bundler',
-                allowImportingTsExtensions: true
-            });
-        }
-        const child = spawn('node', args, { cwd: path.dirname(__filename), env });
+        const args = [testFile];
+        const child = spawn('node', args, { cwd: path.dirname(__filename), env: process.env });
         
         let output = '';
         child.stdout.on('data', (data) => {
@@ -46,10 +35,10 @@ async function runTest(testFile, testName) {
 
 async function runAllTests() {
     const tests = [
-        { file: 'validation-test.mjs', name: 'Validation System Tests' },
+        { file: 'validation-test.cjs', name: 'Validation System Tests' },
         { file: 'app-functionality.test.cjs', name: 'Application Functionality Tests' },
-        { file: 'transformation-applicability.test.mjs', name: 'Transformation Applicability Tests' },
-        { file: 'transformation-apply.test.mjs', name: 'Transformation Apply Tests' }
+        { file: 'transformation-applicability.test.cjs', name: 'Transformation Applicability Tests' },
+        { file: 'transformation-apply.test.cjs', name: 'Transformation Apply Tests' }
     ];
     
     let allPassed = true;
