+++ test/transformation-applicability.test.cjs
@@ -1,11 +1,6 @@
 // Transformation applicability tests
 console.log('=== Transformation Applicability Tests ===\n');
 
-process.env.TS_NODE_COMPILER_OPTIONS = JSON.stringify({ module: 'CommonJS' });
-require('ts-node/register/transpile-only');
-
-const { getApplicableTransformations } = require('../src/utils/transformation.ts');
-
 function runTest(testName, testFn) {
     try {
         const result = testFn();
@@ -100,28 +95,37 @@ const transformations = [
     }
 ];
 
-console.log('🔁 Applicability Matching:');
+const run = async () => {
+    const { getApplicableTransformations } = await import('../src/utils/transformation.ts');
 
-test('Transformation matches external ports of selection', () => {
-    const selection = ['filter-1', 'normalize-1'];
-    const applicable = getApplicableTransformations(nodes, wires, selection, transformations);
-    assertTrue(applicable.some(t => t.name === 'Filter + Normalize -> OptimizedFilter'), 'Expected applicable transform');
-    assertTrue(!applicable.some(t => t.name === 'Mismatched Output'), 'Did not expect mismatched output transform');
-});
+    console.log('🔁 Applicability Matching:');
 
-test('Internal wires are ignored when computing external ports', () => {
-    const selection = ['filter-1', 'normalize-1'];
-    const applicable = getApplicableTransformations(nodes, wires, selection, transformations);
-    assertTrue(!applicable.some(t => t.name === 'Double Input Needed'), 'Internal port should not count as external');
-});
+    test('Transformation matches external ports of selection', () => {
+        const selection = ['filter-1', 'normalize-1'];
+        const applicable = getApplicableTransformations(nodes, wires, selection, transformations);
+        assertTrue(applicable.some(t => t.name === 'Filter + Normalize -> OptimizedFilter'), 'Expected applicable transform');
+        assertTrue(!applicable.some(t => t.name === 'Mismatched Output'), 'Did not expect mismatched output transform');
+    });
 
-test('Unconnected ports are treated as external', () => {
-    const selection = ['lonely-1'];
-    const applicable = getApplicableTransformations(nodes, wires, selection, transformations);
-    assertTrue(applicable.some(t => t.name === 'Filter + Normalize -> OptimizedFilter'), 'Unconnected ports should still match');
-});
+    test('Internal wires are ignored when computing external ports', () => {
+        const selection = ['filter-1', 'normalize-1'];
+        const applicable = getApplicableTransformations(nodes, wires, selection, transformations);
+        assertTrue(!applicable.some(t => t.name === 'Double Input Needed'), 'Internal port should not count as external');
+    });
 
-console.log(`\n=== Test Summary: ${passed}/${total} passed ===`);
-if (passed !== total) {
+    test('Unconnected ports are treated as external', () => {
+        const selection = ['lonely-1'];
+        const applicable = getApplicableTransformations(nodes, wires, selection, transformations);
+        assertTrue(applicable.some(t => t.name === 'Filter + Normalize -> OptimizedFilter'), 'Unconnected ports should still match');
+    });
+
+    console.log(`\n=== Test Summary: ${passed}/${total} passed ===`);
+    if (passed !== total) {
+        process.exitCode = 1;
+    }
+};
+
+run().catch((error) => {
+    console.error(error);
     process.exitCode = 1;
-}
+});
