+++ test/validation-test.cjs
@@ -1,11 +1,6 @@
 // JSON Schema Validation Tests (real schema validation)
 console.log('=== JSON Schema Validation Tests ===\n');
 
-process.env.TS_NODE_COMPILER_OPTIONS = JSON.stringify({ module: 'CommonJS' });
-require('ts-node/register/transpile-only');
-
-const { validateDiagram, validatePalette, validateTransformation } = require('../src/utils/validation.ts');
-
 function runTest(testName, testFn) {
     try {
         const result = testFn();
@@ -130,41 +125,106 @@ const invalidTransformation = {
     replacementNodes: []
 };
 
-console.log('🔎 Diagram Schema Validation:');
-test('Valid diagram passes validation', () => {
-    const errors = validateDiagram(validDiagram);
-    assertTrue(errors.length === 0, 'Expected no validation errors');
-});
+const run = async () => {
+    const { validateDiagram, validatePalette, validateTransformation } = await import('../src/utils/validation.ts');
 
-test('Invalid diagram fails validation', () => {
-    const errors = validateDiagram(invalidDiagram);
-    assertTrue(errors.length > 0, 'Expected validation errors');
-});
+    console.log('🔎 Diagram Schema Validation:');
+    test('Valid diagram passes validation', () => {
+        const errors = validateDiagram(validDiagram);
+        assertTrue(errors.length === 0, 'Expected no validation errors');
+    });
 
-console.log('\n🎨 Palette Schema Validation:');
-test('Valid palette passes validation', () => {
-    const errors = validatePalette(validPalette);
-    assertTrue(errors.length === 0, 'Expected no validation errors');
-});
+    test('Invalid diagram fails validation', () => {
+        const errors = validateDiagram(invalidDiagram);
+        assertTrue(errors.length > 0, 'Expected validation errors');
+    });
 
-test('Invalid palette fails validation', () => {
-    const errors = validatePalette(invalidPalette);
-    assertTrue(errors.length > 0, 'Expected validation errors');
-});
+    console.log('\n🎨 Palette Schema Validation:');
+    test('Valid palette passes validation', () => {
+        const errors = validatePalette(validPalette);
+        assertTrue(errors.length === 0, 'Expected no validation errors');
+    });
 
-console.log('\n🔁 Transformation Schema Validation:');
-test('Valid transformation passes validation', () => {
-    const errors = validateTransformation(validTransformation);
-    assertTrue(errors.length === 0, 'Expected no validation errors');
-});
+    test('Invalid palette fails validation', () => {
+        const errors = validatePalette(invalidPalette);
+        assertTrue(errors.length > 0, 'Expected validation errors');
+    });
 
-test('Invalid transformation fails validation', () => {
-    const errors = validateTransformation(invalidTransformation);
-    assertTrue(errors.length > 0, 'Expected validation errors');
-});
+    console.log('\n🔁 Transformation Schema Validation:');
+    test('Valid transformation passes validation', () => {
+        const errors = validateTransformation(validTransformation);
+        assertTrue(errors.length === 0, 'Expected no validation errors');
+    });
 
-console.log(`\n=== Test Summary: ${passed}/${total} passed ===`);
+    test('Invalid transformation fails validation', () => {
+        const errors = validateTransformation(invalidTransformation);
+        assertTrue(errors.length > 0, 'Expected validation errors');
+    });
 
-if (passed !== total) {
+    console.log('\n🧩 Schema Edge Cases:');
+    test('Empty port arrays are allowed', () => {
+        const diagram = {
+            name: 'Empty Ports',
+            nodes: [
+                {
+                    id: 'node1',
+                    type: 'Empty',
+                    x: 0,
+                    y: 0,
+                    properties: { name: 'Empty', inputs: [], outputs: [] }
+                }
+            ],
+            customNodeDefs: [],
+            wires: []
+        };
+        const errors = validateDiagram(diagram);
+        assertTrue(errors.length === 0, 'Expected empty port arrays to be valid');
+    });
+
+    test('Invalid port name fails validation', () => {
+        const badPalette = [
+            { name: 'BadPort', inputs: [''], outputs: ['out'] }
+        ];
+        const errors = validatePalette(badPalette);
+        assertTrue(errors.length > 0, 'Expected invalid port name error');
+    });
+
+    test('Node properties must be an object', () => {
+        const diagram = {
+            name: 'Bad Properties',
+            nodes: [
+                {
+                    id: 'node1',
+                    type: 'Bad',
+                    x: 0,
+                    y: 0,
+                    properties: 'not-an-object'
+                }
+            ],
+            customNodeDefs: [],
+            wires: []
+        };
+        const errors = validateDiagram(diagram);
+        assertTrue(errors.length > 0, 'Expected invalid properties error');
+    });
+
+    test('Missing diagram name fails validation', () => {
+        const diagram = {
+            nodes: [],
+            customNodeDefs: [],
+            wires: []
+        };
+        const errors = validateDiagram(diagram);
+        assertTrue(errors.length > 0, 'Expected missing name error');
+    });
+
+    console.log(`\n=== Test Summary: ${passed}/${total} passed ===`);
+    if (passed !== total) {
+        process.exitCode = 1;
+    }
+};
+
+run().catch((error) => {
+    console.error(error);
     process.exitCode = 1;
-}
+});
